fix(resources rest): /resources/{resourceGroupName}/{releaseName} hiding /resources/exists/{resourceId} By adding @Produces(APPLICATION_JSON) to /resources/{resourceGroupName}/{releaseName} it was hiding /resources/exists/{resourceId}. Fixed by removing the exist and name rest endpoints and replacing them with the /resource/{id} endpoint.
